cbindfill(tile)	R Documentation

Combine uneven vectors or matrices

Description

Take a sequence of vector or matrix arguments and combine by columns or rows, respectively, regardless of their length

Usage


cbindfill(...)
rbindfill(...)

Arguments

... 	 vectors or matrices to be combined

Details

cbindfill and rbindfill combine a set of vectors of matrices into a single matrix by columns or rows, respectively. Unlike cbind or rbind, the inputs need not have the same number of rows or columns. Instead, the output matrix of cbindfill (rbindfill) has the number of rows (columns) as the longest input. Shorter inputs are filled out with NAs, not recycled.

Value

A matrix combining the ... arguments column-wise or row-wise.

Author(s)

Christopher Adolph <cadolph@u.washington.edu>

See Also

cbind, rbind

Examples


cbindfill(1:5, 1:8, 1:7)
rbindfill(1:5, 1:8, 1:7)

[Package tile version 0.2 Index]