linesTile(tile)	R Documentation

Add a connected line segment to a tile plot

Description

Initializes a simple line graphic. For more advanced features suitable for summarizing inference from a model, use instead lineplot.

This function does no plotting; it only creates a linesTile object to be drawn on one or more plots in a tiled arrangement of plots. To complete the drawing include the object as an input to tile. From tile, it is possible to set further options including plot and axis titles, axis ranges and labels, logged axes, and annotations to the plot.

Usage


linesTile(...)

Arguments

... 	 Any number of arguments given below. Must include exactly one horizontal dimension (x or top) and exactly one vertical dimension (y or right). All inputs should be identified by appropriate tags; i.e., use linesTile(x=myxvar, y=myyvar), not linesTile(myxvar,myyvar)

Details

linesTile simply creates a line or series of connected line segments to add to a tile plot. It serves as the tile equivalent to the base graphics lines or grid graphics linesGrob, and is primarily useful for annotating plots which also use other traces. If you need to draw a large number of disconnected line segments, creating a single trace made by polylinesTile will be much faster than creating a large number of linesTile traces.

The parameters available for constructing a linesTile trace are similar to other trace functions for tile.

trace functions, then layout function

Value

A linesTile object, used only as an input to tile

Author(s)

Christopher Adolph <cadolph@u.washington.edu>

See Also

tile, polylinesTile, lineplot

[Package tile version 0.2 Index]