logitsimev(simcf)	R Documentation

Simulate quantities of interest and confidence intervals for binary logit

Description

Simulate and summarize uncertainty of conditional expected values, first differences and (eventually) other quantities of interest from estimated binary logit models

Usage


logitsimev(x, b, ci=0.95, constant=1)

logitsimfd(xpre, xpost, b, ci=0.95, constant=1)

Arguments

x 	 vector or matrix, counterfactual values of the covariates. Include multiple rows to simulate different counterfactual scenarios

b 	 matrix, simulated parameters, one row per draw from the estimated model, and one column per parameter, including any constants

ci 	 vector, the requested intervals of the simulated quantity of interest to be reported

constant 	 scalar, the column of b containing the model constant, or NA for no constant

xpre 	 vector or matrix, counterfactual initial values of the covariates. Include multiple rows to simulate different counterfactual scenarios; rows must match xpost

xpost 	 vector or matrix, counterfactual final values of the covariates. Include multiple rows to simulate different counterfactual scenarios; rows must match xpre

Details

Given simulated parameters from an estimated logit model, and counterfactual values of the covariates, these functions calculate either the conditional expected value of the response, or the conditional first difference, and confidence intervals around that point estimate.

If the function you used to estimate the model does not provide simulated parameter values, you can draw often them yourself, e.g., using functions such as vcov and mvrnorm in the MASS package, as shown below.

zelig, in the package Zelig, offers similar features for a wide array of models and with automated handling of the simulation process. These functions are offered as a simple alternative for users with simulations already in hand.

Value

Returns a list with three components

pe 	 vector, the point estimate(s) of the requested quantity of interest

lower 	 matrix, the requested lower bounds around the quantity of interest; rows are scenarios, columns are intervals

upper 	 matrix, the requested upper bounds around the quantity of interest; rows are scenarios, columns are intervals

Author(s)

Christopher Adolph <cadolph@u.washington.edu>

See Also

probitsimev, mlogitsimev

[Package simcf version 0.1 Index]