mlogitsimev(simcf)	R Documentation

Simulate quantities of interest and confidence intervals for multinomial logit

Description

Simulate and summarize uncertainty of conditional expected values, first differences and (eventually) other quantities of interest from estimated multinomial logit models

Usage


mlogitsimev(x, b, ci=0.95, constant=1, z=NULL, g=NULL)

mlogitsimfd(xpre, xpost, b, ci=0.95, constant=1,
            zpre=NULL, zpost=NULL, g=NULL)

Arguments

x 	 vector or matrix, counterfactual values of covariates with category-specific parameters. Include multiple rows to simulate different counterfactual scenarios

b 	 3d array, simulated parameters, one row per draw from the estimated model, one column per parameter, including any constants, and one "sheet" for all categories except the reference category.

ci 	 vector, the requested intervals of the simulated quantity of interest to be reported

constant 	 scalar, the column of b containing the model constant, or NA for no constant

z 	 3d array, counterfactual values of covariates specific to the response category. Include one row for each counterfactual scenario, one column for each parameter, and one "sheet" for each category (do not omit any reference categories here)

g 	 3d array, simulated parameters, one row per draw from the estimated model, one column per parameter, and one "sheet" for all categories (do not omit any reference categories here)

xpre 	 vector or matrix, counterfactual initial values of covariates with category-specific parameters. Include multiple rows to simulate different counterfactual scenarios

xpost 	 vector or matrix, counterfactual final values of covariates with category-specific parameters. Include multiple rows to simulate different counterfactual scenarios

zpre 	 3d array, counterfactual initial values of covariates specific to the response category. Include one row for each counterfactual scenario, one column for each parameter, and one "sheet" for each category (do not omit any reference categories here)

zpost 	 3d array, counterfactual final values of covariates specific to the response category. Include one row for each counterfactual scenario, one column for each parameter, and one "sheet" for each category (do not omit any reference categories here)

Details

Given simulated parameters from an estimated multinomial logit model, and counterfactual values of the covariates, calculate either the conditional expected value of the response, or the conditional first difference, and confidence intervals around that point estimate.

If the function you used to estimate the model does not provide simulated parameter values, you can draw often them yourself, e.g., using functions such as vcov and mvrnorm in the MASS package, as shown below.

You must provide either observation-specific (x) or category-specific (z) covariates, or both, and the appropriate parameters (b, g, or both, respectively). Include any observation-and-category-specific covariates in z.

zelig, in the package Zelig, offers similar features for a wide array of models and with automated handling of the simulation process. These functions are offered as a simple alternative for users with simulations already in hand.

Value

Returns a list with three components

pe 	 matrix, the point estimate(s) of the requested quantity of interest; rows are scenarios, and columns are categories

lower 	 array, the requested lower bounds around the quantity of interest; rows are scenarios, columns are intervals, and the third dimension is the category of the response

upper 	 array, the requested upper bounds around the quantity of interest; rows are scenarios, columns are intervals, and the third dimension is the category of the response

Author(s)

Christopher Adolph <cadolph@u.washington.edu>

See Also

logitsimev

Examples


# Multinomial Logistic Regression of alligator food
# See tile package function lineplot for graphical option of this example

data(gator)
require(MASS)
require(nnet)

# Estimate MNL using the nnet library
mlogit.result <- multinom(food ~ size + female, Hess=TRUE)
pe <- mlogit.result$wts[c(6,7,8,10,11,12)]
                                      # point estimates
vc <- solve(mlogit.result$Hess)       # var-cov matrix

# Alternative code for simulations above which calculates CIs to pass on
# to lineplot & tile (also could try Zelig)
sims <- 10000
simbetas <- mvrnorm(sims,pe,vc)       # draw parameters, using MASS::mvrnorm
simb <- array(NA, dim = c(sims,3,2))  # re-arrange simulates to array format
simb[,,1] <- simbetas[,1:3]           #   for MNL simulation
simb[,,2] <- simbetas[,4:6]
sizerange <- seq(1,4,by=0.1)          # range of counterfactual sizes
femalerange <- c(0,1)                 # range of counterfactual sexes

# Create full factorial set of counterfactuals
xhyp <- setfactorial(size = sizerange, female = femalerange)
                                      
# Simulate expected probabilities
mlogit.qoi1 <- mlogitsimev(xhyp,simb,ci=0.67)

print(mlogit.qoi1)

[Package simcf version 0.1 Index]