setfactorial(simcf)	R Documentation

Create a set of counterfactuals using a factorial design

Description

Converts multiple sets of hypothetical covariate values into a single dataframe with all possible combinations of covariate values

Usage


setfactorial(...)

Arguments

... 	 vectors of hypothetical values for each of several covariates, usually provided in tag = value format, where tag matches the name of the original (ie, factual) data

Details

Useful for creating counterfactuals to input into the post-estimation simulation functions of the simcf package or into whatif in the WhatIf package.

Value

A dataframe with columns corresponding to the provided variables, and rows containing every possible combination of the given covariate values

Author(s)

Christopher Adolph <cadolph@u.washington.edu>

Examples


xhyp <- c(1,2,3)
whyp <- c(TRUE, FALSE)
zhyp <- c("a","b","c")
setfactorial(x = xhyp, w = whyp, z = zhyp)

[Package simcf version 0.1 Index]