tileDraw(tile)	R Documentation

(Re-)draw a tile graphical object

Description

This internal function draws a tile graphical object to the current device. It is mainly of use of an internal function; most users should simply use tile to create and save tile graphics to a file.

Usage


tileDraw(grob, defaultlayer=10)

Arguments

grob 	 A tile or grid graphical object; e.g., the object in the slot grob in the value returned by tile

defaultlayer 	 The layer to draw any sub-elements of grob. Layers with higher numbers are drawn first, then lower numbers are drawn on top.

Details

This internal function is provided for completeness, and is not necessary to create tile graphics with detailed annotations, which can be done through careful construction of a single call to tile.

However, if users wish to add to or modify an existing tile grob, they may do so by directly editing the grob returned by tile, then sending it to this function to place it on the current device. Care should be taken to set up the current device to have appropriate height, width, and pointsize.

Value

tileDraw is mainly called for the side-effect of drawing a tile grob. It invisibly returns the grob after drawing.

Author(s)

Christopher Adolph <cadolph@u.washington.edu>

See Also

tile

[Package tile version 0.2 Index]